ré-organise les fichiers associés aux listes et filtres d'indicateurs
