+++ app.territoiresentransitions.react/src/app/pages/collectivite/Indicateurs/lists/IndicateurChartsGrid.tsx
@@ -1,9 +1,12 @@
 import {Link} from 'react-router-dom';
 import {useCollectiviteId} from 'core-logic/hooks/params';
 import {useIntersectionObserver} from 'utils/useIntersectionObserver';
-import IndicateurChart from './charts/IndicateurChart';
-import {TIndicateurChartProps} from './charts/types';
-import {TIndicateurDefinition, TIndicateurReferentielDefinition} from './types';
+import IndicateurChart from '../charts/IndicateurChart';
+import {TIndicateurChartProps} from '../charts/types';
+import {
+  TIndicateurDefinition,
+  TIndicateurReferentielDefinition,
+} from '../types';
 import {makeCollectiviteIndicateursUrl} from 'app/paths';
 
 type TIndicateurChartsGridProps = {
